Adds conditional compilation for IDL builds
Adds conditional compilation to handle different build scenarios,
specifically when building IDLs (Interface Definition Languages).

This ensures correct program compilation when the `idl-build` feature
is enabled and prevents entrypoint generation in that case, while also
providing a global allocator when the `idl-build` feature is enabled
and the `no-entrypoint` feature is not.